# doc/openapi-3.yaml
openapi: 3.0.1
info:
  title: WebMarketREST
  contact:
    email: gea.viozzi@student.univaq.it
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '0.1'
...
servers:
  - url: http://localhost:8080/WebMarketREST
...
paths:
  /rest/richieste:
    post:
      operationId: addRequest
      description: adds a new request
      tags:
        - Richieste collection
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaOrdine'
      responses:
        '201':
          $ref: '#/components/responses/SuccessfulInsert'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/GeneralError'
  /rest/richieste/in_attesa:
    get:
      operationId: getRichiesteInAttesa
      description: >-
        trova tutte le richieste non prese in carico da nessun tecnico
      tags:
        - Richieste collection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista di richieste in attesa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RichiestaOrdineBase'
              example: 
               id: 5,
               note: "Acquisto sedia ergonomica"
               stato: "IN_ATTESA"
               codiceRichiesta: "52kSv6aSOQ"
               data: "2024-09-02"
               utente:
                username: "geaviozzi"
                email: "geaviozzi@icloud.com"
               categoria:
                id: 10
                nome: "Sedie"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/GeneralError'
  /rest/richieste/in_corso:
    get:
      operationId: getRichiesteInCorso
      description: >-
        trova le richieste in corso di un determinato utente
      tags:
        - Events collection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: richieste in corso di un determinato utente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RichiestaOrdineBase'
        '401':
          $ref: '#/components/responses/UnauthorizedError'        
        '500':
          $ref: '#/components/responses/GeneralError'
  /rest/richieste/gestite_da_tecnico/{uid}:
    parameters:
      - $ref: '#/components/parameters/TecnicoUID'
    get:
      operationId: getRichiesteGestiteDa
      description: trova le richieste gestite da un determinato tecnico
      tags:
        - Richieste Collection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Richieste info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RichiestaOrdineBase'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/GeneralError'
  /rest/richieste/{uid}:
    parameters:
      - $ref: '#/components/parameters/RichiestaUID'
    get:
      operationId: getRichiesta
      summary: Ottieni i dettagli di una Richiesta specifica
      description: Restituisce i dettagli di una singola `RichiestaOrdine` in base all'`idrichiesta` fornito.
      tags:
        - Richiesta
        - Richiesta properties
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Richiesta info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RichiestaOrdineBase'
              example:
                id: 21
                note: "Vorrei un mobile da studio"
                stato: "PRESA_IN_CARICO"
                codiceRichiesta: "56ghjkl90"
                utente:
                  username: "samanta"
                  email: "samanta_dis@hotmail.com"
                tecnico:
                  username: "tecnicoS"
                  email: "tecnicoS@hotmail.com"
                categoria:
                  id: 8
                  nome: "Mobili"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/GeneralError'
    delete:
      operationId: deleteRichiesta
      description: elimana la richiesta
      tags:
        - Richiesta
      security:
        - bearerAuth: []
      
      responses:
        '204':
          description: Richiesta eliminata con successo
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/GeneralError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  responses:
    SuccessfulInsert:
     description: Elemento creato
     headers:
      location:
        schema:
         type: string
         format: uri
     content:
      application/json:
       schema:
        type: string
        format: uri
    BadRequest:
      description: Errore nella richiesta 
    GeneralError:
      description: Errore Interno del Server
    UnauthorizedError:
      description: Token mancante o non trovato
    NotFound:
     description: Elemento non trovato
     content:
      text/plain:
       schema:
        type: string
        example: Not found
  parameters:
    TecnicoUID:
      name: uid
      in: path
      required: true
      schema:
       $ref: '#/components/schemas/uid'
    RichiestaUID:
      name: uid
      in: path
      required: true
      schema:
       $ref: '#/components/schemas/uid'
...
  schemas:
    uid:
      type: integer
      pattern: "[0-9]+"
    Utente:
      type: object
      properties:
        uid: 
          $ref: '#/components/schemas/uid'
        username:
          type: string
        email:
          type: string
          format: e-mail
        password:
          type: string
        tipologiaUtente:
          enum:
            - ORDINANTE
            - TECNICO
      required:
        - uid
        - username
        - email
        - password
        - tipologiaUtente
      additionalProperties: false
    PropostaAcquisto:
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/uid'
        produttore:
          type: string
        prodotto:
          type: string
        codiceProdotto:
          type: string
        codice:
          type: string
        prezzo:
          type: number
        url:
          type: string
        note:
          type: string
        stato:
          enum:
            - ACCETTATO
            - RIFIUTATO
            - IN_ATTESA
            - ORDINATO
        data:
          type: string
          format: date-time
        motivazione:
          type: string
        richiestaOrdine:
          $ref: '#/components/schemas/RichiestaOrdine'
      required:
        - uid
        - produttore
        - prodotto
        - codiceProdotto
        - codice
        - prezzo
        - url
        - stato
        - data
        - richiestaOrdine
      additionalProperties: false
    RichiestaOrdineBase:
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/uid'
        note:
          type: string
        stato:
          enum:
            - IN_ATTESA
            - PRESA_IN_CARICO
            - ORDINATA
            - RISOLTA
        data:
          type: string
          format: date-time
        codiceRichiesta:
          type: string
        tecnico:
          $ref: '#/components/schemas/Utente'
        utente:
          $ref: '#/components/schemas/Utente'
        categoria:
          $ref: '#/components/schemas/Categoria'
    RichiestaOrdine:
      type: object
      properties:
        richiesta:
          $ref: '#/components/schemas/RichiestaOrdineBase'
        proposte:
          type: array
          description: Lista di proposte associate alla Richiesta in questione
          items:
            $ref: '#/components/schemas/PropostaAcquisto'
        caratteristiche:             
          type: object
          description: Una mappa di caratteristiche scelte dall'ordinante, associate alla Richiesta specifica.
          additionalProperties:
            type: string 
      required:
          - uid
          - stato
          - data
      additionalProperties: false
...
    Categoria:
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/uid'
        nome:
          type: string
        caratteristiche:
          type: array
          items:
            $ref: '#/components/schemas/Caratteristiche'
      required:
          - uid
          - nome    
      additionalProperties: false
    Caratteristiche:
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/uid'
        nome: 
          type: string
        categoria:
          $ref: '#/components/schemas/Categoria'
      required:
          - uid
          - nome
          - categoria